Move commands to separate files
