+++ palang-cli/src/commands/mod.rs
@@ -1,0 +1,6 @@
+pub mod compile;
+pub mod run;
+pub mod serve;
+pub mod status;
+pub mod projects;
+pub mod profiles;
